feat: add view edit update invoice.

+++ database/migrations/2026_02_19_182559_create_invoices_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->foreignId('debitor_id')->constrained()->cascadeOnDelete();
             $table->foreignId('debitor_site_id')->nullable()->constrained()->nullOnDelete();
 
-            $table->decimal('total_qty', 10, 2)->default(0);
+            $table->bigInteger('total_qty')->default(0);
             $table->decimal('total_amount', 15, 2)->nullable()->default(0);
             $table->decimal('total_a_amount', 15, 2)->nullable()->default(0);
